add css and semantic html to recipe display and top bar nav
