+++ client/src/App.css
@@ -37,9 +37,80 @@
   }
 }
 
+#NavBar{
+  background-color: #2BB852;
+  color: white;
+  position: fixed;
+  height: 90px;
+  top: 0;
+  width: 100%;
+  display: flex;
+  flex-direction: row;
+  
+}
+
+#navbar-icon{
+  height: 100px;
+  width:100px,
+}
+
+#navbar-options{
+  display: flex;
+  flex-direction: row;
+  align-items: flex-end
+}
+
+#navbar-options a{
+  color: white;
+  float: left;
+  display: block;
+  text-align: center;
+  padding: 5px 15px;
+  font-weight: bold;
+  text-decoration: none
+  
+}
+
+.allRecipesContainer{
+  display: flex;
+  flex-flow: row wrap;
+  justify-content: space-around;
+  gap: 60px 50px
+
+}
+.singleRecipeContainer{
+  border-style: solid;
+  border-color: #c0c0c0;
+}
+
+.singleRecipeContainer a{
+   color: rgb(77, 77, 77);
+   text-decoration: none;
+   display: block;
+   text-align: center;
+   font-weight: bold;
+}
+
+.singleRecipeDisplayIcon{
+  height: 300px;
+  width: 300px
+}
+
 #recipeImg{
-max-height: 500px;
-max-width: 500px
+  max-height: 500px;
+  max-width: 500px
+}
+
+.imagesContainer{
+  display:flex;
+  flex-flow: row wrap;
+  align-content: flex-start;
+  gap: 50px 30px ;
+}
+
+.imgContainer{
+  display:flex;
+  flex-direction: column;
 }
 
 .img-container{
@@ -52,7 +123,13 @@ max-width: 500px
   display:flex;
   flex-direction: row;
   justify-content: space-around;
+}
 
+.recipebuilder_availableImages{
+  display:flex;
+  flex-flow: row wrap;
+  justify-content: space-around;
+  gap: 10px
 }
 
 .basket{
@@ -61,7 +138,6 @@ max-width: 500px
   width:500px;
   display:flex;
   flex-direction: row;
-
 }
 
 .recipeSegment{
@@ -86,8 +162,12 @@ max-width: 500px
   max-width: 500px;
   text-align: center;
   background-color: #c0c0c0;
-    position:fixed;
-    top:0;
-    width:100%;
-    z-index:100;
+  position:fixed;
+  top:0;
+  width:100%;
+  z-index:100;
+}
+
+.main-content-container {
+  margin-top: 100px; /* Add a top margin to avoid content overlay */
 }
