+++ client/src/App.js
@@ -1,22 +1,19 @@
 
 import './App.css';
-import { Link } from "react-router-dom";
-
+import Navbar from './Navigation/Navbar'
+import AppRoutes from './Routes';
 
 function App() {
   return (
-    <div>
-      <div>
-        <h1>Recipe Reader!</h1>
-        <Link to="/upload">Upload New Images</Link> |{" "}
-        <Link to="/recipes">Your Recipes</Link> |{" "}
-        <Link to="/files">Your Uploaded Files</Link> | {" "}
-        <Link to="/newRecipeImages">Create New Recipe</Link> 
-      </div>
-        <div>
-        </div>
-  </div>
-
+    <>
+      <header>
+        <title>Recipe Reader</title>
+        <Navbar/>
+      </header>
+      <main>
+        <AppRoutes className = 'main-content'/>
+      </main>
+  </>
   );
 }
