+++ client/src/Home.js
@@ -1,0 +1,8 @@
+
+export default function Home(){
+    return(
+        <>
+        <h1>Welcome To Recipe Reader</h1>
+        </>
+    )
+}
