+++ client/src/Navigation/Navbar.js
@@ -1,0 +1,23 @@
+import { Link } from "react-router-dom"
+
+export default function NavBar(){
+    return(
+      <>
+        <nav id = "NavBar">
+          <div>
+            <Link to ="/">
+              <img 
+                id = "navbar-icon"
+                alt = "defeaultImg"
+                src = "RRIcon.png"/>
+            </Link>
+          </div>
+          <div id = "navbar-options">
+            <Link to="/recipes">Recipes</Link> 
+            <Link to="/files">Images</Link> 
+            <Link to="/newRecipeImages">Create Recipe</Link> 
+          </div>
+        </nav>
+      </>)
+
+}
