+++ client/src/RecipeBuilder/RecipeBuilder.js
@@ -4,7 +4,7 @@ import React, { useState, useEffect, createContext } from 'react'
 import { HTML5Backend } from 'react-dnd-html5-backend'
 import UserImages from "./UserImages"
 import Basket from "./DropBasket"
-import apiService from '../apiService';
+import apiService from '../Utilities/apiService';
 import {Link} from "react-router-dom"
 
 export const Context = React.createContext()
