+++ client/src/RecipeEditor.js
@@ -1,5 +1,5 @@
 import {Editor, EditorState, ContentState} from 'draft-js'
-import apiService from "./apiService";
+import apiService from "./Utilities/apiService";
 import React from 'react'
