+++ client/src/Utilities/helperFunctions.js
